feat(ens): Move normalizeEnsHash function to ens-utils.js for better organization
